[#24] Stepper open loop tested with stepper motor, old budge functions use new helper functions but don't work, but software interrupt stepper control works. Not a big concern since old functions use blocking loops which are undesirable.

+++ robot/rover/ArmDriverUnit/Test/StepperOpenLoop/RobotPID.h
@@ -5,7 +5,7 @@ class RobotPID {
   public:
     //temp stuff for stepper open loop control
     int openLoopDir;
-    int openLoopError;
+    float openLoopError;
     int numSteps;
     volatile int stepCount;
